app/requirements_portable.txt
selenium>=4.15.0
python-dotenv>=1.0.0
requests>=2.25.0
InquirerPy>=0.3.4
colorama>=0.4.6